fix(arm64) Fix tests so they work on arm64 as well
This is mostly fixing a bug with our current tests, where in the old version of
Clickhouse the events are not being merged before the tests run. In the new
version of Clickhouse the events are merged before the tests, and since the
fixtures all use the same span_id, they are actually only sending one event.