feat(admin) Add authorization provider in snuba admin (#2301)
The admin relies on a proxy to perform authentication and authorization.
This is not great for two reasons:

it would by default allow everybody if the proxy was misconfigured while failing open (blocking everybody) would be safer.
it does not allow to decide which features to show depending on roles.
This adds an abstraction to performa authorization before each server request on the admin UI.
This abstraction has multiple implementations for multiple authorization providers. The default is NOOP and allows everything through. The next to be implemented is IAP, then we can add one for basic HTTP auth useful on prem.