feat(snql) Support higher order functions in SnQL (#2333)
* feat(snql) Support higher order functions in SnQL

SnQL now supports using lambdas in its clauses. Lambdas are defined as a tuple
of identifiers, followed by an arrow, followed by a function that may or may not
use those identifiers. An identifier in SnQL is a set of characters enclosed in
backticks. The distinction to use backticks was to make parsing and validating
easier, since the parser doesn't have to try and determine if a string is a
column or an identifier.

This feature set is slightly more limited than what is available in Clickhouse.
The major difference is that a lambda can only be used as a parameter in a
function, and the transformation of a lambda must always be a function itself.
So (`x`) -> `x` is not a valid lambda.

This change makes using higher order functions such as arrayMap possible, since
those functions required a lambda to be defined to use them.