Enforce uniq in HAVING and SELECT (#2339)
Context

This is a hacky query processor put in for the purpose of
mitigating clickhouse crashes. When a query is sent to clickhouse
which has a uniq in the HAVING clause but not in the SELECT
clause, the query node will crash. This may be fixed with future versions
of clickhouse but we have not upgraded yet. This exists as a check so that we
return an exception to the user but don't crash the clickhouse query node

Solution

    add query processor to catch this case, make it throw an exception
    Add query processor to transaction and error storage