feat(admin) Expose tracing endpoint in the UI
Expose the tracing endpoint with a UI. It follows the same format as the
Clickhouse system queries form, but it can run any arbitrary query. The result
is the raw trace logs that Clickhouse produces when running the query.

Also create a common library for displaying query results that is shared between
components.