feat(suspect-spans): Add array has processor (#2348)
SnQL only allows for conditions of the form `LHS OPERATOR RHS`. This means that
conditions like `has(column, value)` are currently illegal, and has to be
written as `has(column, value) = 1` to satisfy this constraint. This causes
problems with Clickhouse when it's not able to use the bloomfilter index for
these conditions. This change rewrites such conditions to remove the surrounding
`... = 1` so that Clickhouse is able to properly leverage the bloomfilter index.