perf(subscriptions): Record separate metrics for encoding and finding tasks
Record a separate metric for building tasks and encoding tasks so we can get an idea
of where we are spending more time.

Also includes a small change to DelegateTaskBiulder.get_task() so that it only
gets the jittered or immediate task as required instead of building both and
throwing one away.