perf(subscriptions): Add a bunch of metrics to SubscriptionScheduler.find()
This function seems to be taking a really long time relative to other parts
of the scheduler. This change adds a bunch of timings to this function to see
if we can isolate the issue further.

Also cleans up a couple metrics we were previously recording but don't need anymore.