Retry in case of concurrency limit (#2356)
ClickhousePool.execute_robust Is supposed to execute some retries with backoff in case Clickhouse throttles it.
It actually does not happen because it catches the wrong exception.
It catches ServerException instead of ClickhouseError since Whatever error comes from Clickhouse has already been wrapped in ClickhouseError by the execute method.

This means that, as of today, if a subscription query gets throttled by clickhouse no retry is performed and the consumer restarts, this introduces delay most of the times for no reasons. Throttling is generally temporary and lasts for very few seconds.

This PR catches the right exception and introduce a linear backoff between retries.