feat(admin): Add a new collapse component
This adds a collapse UI component to be used in the admin tool.
This is designed to be used to render query history in the system
queries and tracing tool, and replace the table within a table
layout that we have there now. It is not being used anywhere
yet and will be integrated in the future.