fix(tracing): Fix querying the query node in the tracing tool (#2372)
The is_valid_node function previously only checked if the node
being queried was a valid local node since we only currently query
local nodes for system qeuries. Now we need to allow the
query node as well since this is what the tracing tool uses.

This fixes a bug introduced in https://github.com/getsentry/snuba/pull/2366
when we switched over to use the readonly user.