feat: Remove old way of querying canned system queries
We have not been using this old way of querying system tables
for a while, so this removes support for passing `query_name`. The new way
we are building here (https://github.com/getsentry/snuba/pull/2374)
will invole passing the SQL statement from the UI (same as a regular
system query) so just passing a `query_name` to be mapped to a sql statement
by the backend will not be needed.