feat: Share ClickHouse connections to the same nodes between clusters (#2359)
Previously ClickHouse connection pools were cached in the ClickhouseCluster class so they would be reused when attempting to access a ClickHouse node via a cluster. However if the same nodes are present in multiple clusters (as is the case with the way Sentry's query nodes are currently configured) we would create a separate and unnecessary connection pool to those same nodes from each cluster.

This change addresses that issue by moving the connection cache outside of the ClickHouseCluster class and reusing the same cache for all clusters.