ref: Small refactor of admin UI (#2377)
https://github.com/getsentry/snuba/pull/2375 split the UI for tracing
and system queries so they no longer use the same components. This goes
a step further and:
- Moves the query display under the system_queries module since no one
else is using it now
- Splits the executeQuery function in the API client code into separate executeSystemQuery
and executeTracingQuery methods with distinct request and result types. These 2 APIs
are diverging so we don't want to keep typing them in the same way.

This PR only contains refactoring, no functionality or visuals have changed.