fix(admin): Prevent form post (#2386)
Sometimes when you run a system query it just clears the page instead
of properly displaying the results. This is due to the default browser
action associated with posting the form and redirecting to the next page.
Since this is a single page app and we don't want to go anywhere just
suppress the default action with event.preventDefault().