feat(dev): Mock data for subscriptions (#2390)
I needed to build this as part of load testing the scheduler
in https://github.com/getsentry/snuba/pull/2387 and although most
of those other changes will be thrown away, I think this part may be
worth keeping. I wanted an easy way for everyone to be able to get
mock subscription data into Snuba for dev and load testing and right
now it's a bit of a hassle.