fix(consumer): Allow commit log topic to be passed to multistorage consumer (#2393)
While modifying the production config to change transactions consumer to
be multi storage consumer, I realized that the multi storage consumer does
not accept a commit log topic. That makes switching from single storage to
multi storage and vice versa prone to errors.

This PR addresses that by adding a commit log topic option which can be
passed to the multi storage consumer and have parity between the different
options in both consumers.