feat(ingestion): Allow for failures during writes (#2395)
When we introduce a new storage with multi storage consumer, its a good
idea to allow failures during writes since we are not too sure if
everything is in place for end to end writes. This is important
because we don't want commits to kafka to stop when bringing in a new
cluster.

The change introduces an option per storage which allows for failures
during writes.