perf(subscriptions): Avoid slow delegate task builder where possible
The delegate task builder is much slower than using the immediate
or jittered one since it has to check which mode to apply for
every subscription within the tick individually. This change uses
the immediate or jittered task builder directly for the whole tick
if we are not in a transition mode. This is checked at the start
of each tick.