fix(cleanup) Fix rounding error in cleanup script (#2394)
Since the cleanup script wasn't rounding to midnight, it was possible to have a
partition that was still in the retention period, but counted as being out
the script wasn't rounding it's values to midnight.

For example, the script runs at 5 minutes past midnight. It finds a partition
with a last part that is only 2 minutes past midnight. Before, since the
partition date is less than the script date, the partition would be deleted. Now
everything is rounded to midnight so the script will keep the partition.