fix(admin): Fix the system query tool (#2414)
None of the system queries work currently because we are attempting to run them
with readonly=2 but we have only configured readonly=1 for this user.

Since system queries do not actually need the expanded permissions that come with
readonly=2 (ability to change settings queries) just run them with them readonly=1
instead. Tracing still runs with 2.