ref(subscriptions): Remove subscription "type" field (#2403)
Whilst we were transitioning from the legacy format to SnQL for subscriptions
we had to support two formats. Now we only have SnQL subscriptions we can
remove the "type" field from subscription data. This should be safe to do
as we are no longer actually relying on this value for anything.