fix(storages): Allow storage definitions in conf which do not exist (#2417)
Adding a storage today is like a chicken and egg problem. If the code
dealing with a new storage is introduced first and the settings in prod
do not exist, then because of implicit imports its possible that an
exception could be raised. If the prod settings for new storage
are introduced first then because of validation of settings there would
be an exception since the storage mentioned is not in the code.

This PR tries to break this cycle by allowing storages to be mentioned
in the settings without the corresponding storage to be defined in
code.