fix: Avoid warning on duplicate aliases
There are checks to make sure that different expressions can't have the same
alias, so the only way to hit the warning case when building the response is
an edge case where the same expression has two different aliases. Currently
in that case we would log a warning and only return one of the columns in
the response. Instead, don't log a warning and return both columns with
the same values.