feat(errorsv2) Similarity score algorithm to compare results from multiple Clickhouse versions (#2426)
This implements the callback function for the PipelineDelegator. The callback has access to both the primary and secondary query results and tries to calculate a similarity score between the two results.

The idea is to divide each row of the results in a key (strings, UUID, datetime), which are generally what the user put in the GROUP BY, and the numeric fields, which are the results of aggregations. (For simplicity this only relies on the data type to decide what is key and what is value).
If two results are very similar, the keys should match and the values for matching keys should be close to each other.
So the score is the average similarity score for matching rows. Each pair of matching rows (same key between the two results) has a similarity score.
The similarity score of a pair of matching rows is calculated as the average of the ratios of matching columns between the two rows.

Any pair of results that do not share the same schema (same columns) get a score of 0.0.