ref: Change QueryException to show inner error
If a QueryException was `str`, it would json encode the extra_data and no
context on what the inner error was. This isn't a big deal locally, but when
looking at these errors in Sentry, the exception is buried in the inner errors
and makes it harder to diagnose and understand errors. I also think it is
causing Sentry to group errors that it shouldn't.

Add the inner exception message to the `str` of the QueryException so it is
easily readable in Sentry and makes debugging easier.