Add read path for upgraded transactions (#2454)
Basically copy-paste of #2425 but for transactions.

TRANSACTIONS_UPGRADE_BEGINNING_OF_TIME was determined by looking at the first insert query in datadog.

Tests are copy and pasted from the errors one. If this was less urgent I'd be more clever and reuse that code but since we only need to do this for three entities I think it's ok.