feat(pipeline delegator) Split rate limiter namespace (#2448)
In order to rollout to rollout a large amount of queries to errors_v2 we need to split the rate limiter name spaces.
If we do not do that we would double count and reject queries that should be accepted.
The resources to rate limit are disjoint (separate clusters) so there is no issue in having a double quota.

The way to implement this is to leverage the PipelineDelegator so that it provides a a separate instance of RequestSettings to each of the two pipelines to be executed.
This allows us to still make each query pipeline independent and unaware of the pipeline delegator. Keeping each pipeline independent reduces the complexity of the design.