perf(quota) Make max thread configurable per referrer and project (#2424)
Configuring the rate limiter works to protect the system though it does not help when we have continuous moderate load that is far from being limited by concurrency. By having low concurrency every query uses a large amount of threads.
This is ok for the UI for a short period of time, but it is a problem when it comes from script that hit the api continuously.
This allows to manually set the max thread config setting by runtime config and reduce quota for specific referrer/projects.

A further improvement would be to support a proper quota system where the max_thread field is calculated based on the past behavior of the project and adapted to enforce a quota over a period of time.