feat(subscriptions): Commit offsets even if there are no subscriptions registered (#2463)
Previously the scheduler would only commit offsets if there were subscriptions registered
for that partition. This is problematic for subscriptions on new entities/datasets for
which there will be no (or very few) subscriptions to begin with.

This change fixes that by committing for any tick where `should_commit=True` and no
subscriptions are scheduled.