perf(metrics) - consolidate metrics materialized views (#2469)
Consolidate materialized views for writing metrics raw data to aggregate tables. This reduces the number of distinct writes to the aggregate tables and hopefully improves performance.

When deployed and executed these migrations will not take immediate effect. They will be put in place when a settings change is deployed to set materialization_version = 2.

Once we've switched that over and verified that all new rows being written are at 2, we should drop the at that point non-functional materialized views that act on materialization_version = 0