feat(subscriptions): Add utility classes for subscription verifier
Adds a decoder for subscription results and a consumer for consuming
subscription result topics.

The new consumer and decoder are not yet being used, these pieces will be
combined later as part of a "subscriptions verifier". The subscription verifier's
job is to ensure the old and new subscriptions pipeline are producing equivalent
results. It will consist of a synchronized consumer that consumes from two result
topics (using this SubscriptionResultConsumer) and processing strategies that will
diff and report on the results.

All verifier code including this is intended to be temporary and will be removed
together with the old subscription pipeline once we have finished the migration.