ref(shutdown) Add a metric to track incoming queries after shutdown (#2455)
There appears to be a problem where queries are failing that shouldn't be
failing. One theory for this is that queries are being queued waiting for cached
results, and then when the container shuts down those queries are marked as
failed. However we have a 40 second wait for all queries to clear before the
container shuts down. So the only way queries could be waiting for the cache is
if they were sent to the pod after it was marked as shutting down. This metric
will track that case.