perf(consumer): Skip unnecessary processing for multistorage consumers
With the introduction of the v2 storages and changing the production consumer
to multistorage consumer we have had to increase the number of replicas
since we are basically doubling the amount of work that needs to be done.
But the bytes of data which gets written to the existing and v2 storages
are the same since they have the same processors.

This PR introduces a notion of equivalence of storages for writes and if
during processing of messages finds that an equivalent storage has
already been processed, uses its processed value rather than doing the
same work again.