feat(subscriptions): Add synchronized consumer for verifier
The synchronized consumer consist of two SubscriptionResultConsumer
instances that consume from the old and new result topic simultaneously.
It's job is to ensure that we make progress on each topic at roughly the
same pace based on the scheduled subscription timestamp.

It does not attempt to keep track of the progress of individual partitions
on each result topic, that will be handled by the processing strategy later.