fix(consumers): Remove deepcopy from multi storage consumers (#2491)
We recently added the deepcopy in the process_message_multistorage api since the transaction processor was modifying the message in place. Doing a deepcopy would ensure that functionality would work as expected. But it has a side effect of affecting the consumers performance since copying is a heavy operation.

Fix the transaction processor so that the operation where it needs to modify the context does the deepcopy. That way other consumers are not affected.