feat: Add a default value for stats_collection_freq_ms (#2495)
We don't necessarily want to set the config individually for each consumer group
for how often stats are collected, as it pollutes our config and it is easier to
use reasonable default values across all consumers. This change adds a default config
`stats_collection_freq_ms`, which is applied if a config is not specified for the
specific consumer group.