feat(delegate_pipeline) Split cache in partitions per cluster (#2502)
The original issue that motivated this PR is that, during the Clickhouse upgrade, we were double reading from both clusters via the PipelineDelegator. Unfortunately, the query strings were identical because the schema on both storages is identical.
Being the query identical the cache key was the same and we were actually running most of the times only one of the two as the second would be a duplicate of the first.

Overall we could run into identical queries on multiple clusters. While, on a single cluster, identical queries are actually duplicate. So this PR partitions the cache per cluster to avoid these duplications.

This implicitly makes the cache a concept tied to the cluster. Which is de facto true in this implementation as the cache is applied right before rate limiting and sending the query to clickhouse. If we moved the cache to the logical level, of course this design would change and partitioning would probably be per dataset (if any partitioning will be necessary at that point).

This PR:

adds the property to the cluster configuration. If the cache partition is not defined we default to a generic one.
initialize separate RedisCache objects in a thread safe way