fix(transactions): Remove flattened columns (#2506)
The transactions processor currently defines the contexts flattened and
tag flattened columns which are not filled in by the consumers
and not queried by the api's. These columns have a TTL expiry of 1 day
which causes Clickhouse to spend merge TTL on something which we don't
use. Hence removing these columns.