perf(consumer) Prepare to remove the latency metric at every message (#2514)
We are logging a latency metric between clickhouse write an message timestamp at every message. This seems to contribute to up to 1 second latency for every batch, and likely it is the cause of some hard to explain backlog events.

The idea is to only log one latency per batch, as our SLO should be based on the max latency.
Before replacing the existing metric, this PR starts by recording the max latency and the avg latency per batch to compare how consistent is with the current metric. If it is close enough, we can replace the existing one.