Fix settings delegator (#2513)
The current RateLimiterDelegate class has an issue: it replicates the state of the class it delegates to in order to keep the list of transformed rate limiter settings.
This is not a great practice as it assumes the behavior of the delegate class is compliant with HTTPRequestSettings, thus breaking its abstraction.
SubscriptionRequestSettings behave differently: it returns nothing as rate limiter config no matter what you add to it. So in practice RateLimiterDelegate does not respect its behavior because it overrides this behavior.

This has two negative consequences:

functionally RateLimiterDelegate is not compliant with what SubscriptionsRequestSettings thus adding a rate limiter on subscriptions.
In terms of performance it triggers the rate limiter infrastructure on subscriptions, which slows down the pipeline (and required us to scale out the production subscriptions consumer).
This changes the approach. RateLimiterDelegate does not store anything anymore. It fully delegates the state to the delegate class and , instead, it transforms the RateLimitParameters instances on the fly when returning them. This makes the behavior compatible with SubscriptionsRequestSettings (see added test).