fix(subscriptions): Fix committing of offsets in scheduler (#2500)
There were a number of issues with the way committing offsets were
being handled in the subscriptions scheduler previously. This resulted
in offsets not actually being committed when they should be.

- The ProvideCommitStrategy was attempting to set the "committable" status
based on the partition with the earliest timestamp (based on the original topic).
This is wrong, it should actually be the partition with the earliest offset
on the commit log topic that can be safely committed.
- The strategy was attempting to mark all messages as either "committable" or
not. This is not the right approach as it a later message cannot necessarily
be immediately committed, it only makes an earlier one safe to commit. Now
we are passing the offset_to_commit (which is always an earlier one) through
the pipeline together with the actual tick message.
- A bug in ScheduledSubscriptionQueue.peek() was actually causing offsets
to be committed even if they were never marked "committable" to begin with,
resulting in the issue not being discovered earlier as long as there are
any subscriptions that are scheduled.