Increase partition size on metrics raw tabl (#2524)
Daily partition is appealing if most data is concentrated in the latest partition. Unfortunately it seems the content of an INSERT is more evenly divided across partitions. This generates one part per partition almost at every batch, which creates 5 parts every time increasing the number of inactive parts instead of reducing

So this pr makes the partitioning key more coarse. We try a 3 days partition.
This PR is almost identical to
#2504
#2507
except for the partition key and the table names.