feat(optimize) Ensure the optimize job does not go past 11 pm (#2535)
We saw a few times the optimize job being restarted by Kubernetes after a few hours.
That job is stateless so it restarts from scratch. If k8s restarts it too late, the job may end so late to overlap with the next day instance. This puts unnecessary load on the DB.

If we are about to start a new instance there is no point in finishing the previous one, so this adds a cutoff time to ensure the job does not go past a certain time of the dat provided in settings.

The proper fix would be to make the job stateful so that it picks up from where it left when restarted, though that is a large refactoring. This is at least a safeguard to prevent the overlap.