test: Fix typing in tests and run mypy on tests in CI
The goal of this change is to reduce the likelihood of
introducing new tests that are improperly typed or introducing
new code that breaks tests that were previously typed correctly.

Mypy is now run on all tests by default in CI. Only those provided
on the exclude list will be excluded.