Destroy the stateful consumer (#2566)
This consumer was built in order to ensure we would load a Change Data Capture snapshot from postgres consistently without possibility of double writes on Clickhouse.
We never used it. It probably does not work and the way we store CDC data in Clickhouse does not really require avoiding double writes.
So I think we can just remove it entirely and that reduces the reasons for having a consumer builder and makes the consumer building process simpler.