perf(optimize): Allow running 2 clcikhouse optimize queries concurently (#2565)
The optimize jobs have started taking longer to complete. Sometimes
they complete around noon PST. Which means that the job is running
even during peak traffic time.

To help with this, we can try running optimize on 2 different partitions
simultaneously. This should be ok to do since we start the optimize
job at 6 PM PST when the traffic volume is relatively low.