feat(consumers): Support the --cooperative-rebalancing option
The consumers, multistorage-consumers, subscriptions consumer and
subscriptions-executor now support passing a --cooperative-rebalancing
flag which, if passed, will use the "cooperative-sticky" partition
assignment strategy.

Note that cooperative and non-cooperative strategies cannot be
combined in the same group so this cannot be deployed with a rolling
strategy. When enabling cooperative rebalancing for the first time,
all old consumers in the group should be stopped before restarting.