feat(subscriptions): Add basic verifier (#2542)
Add a subscriptions verifier process that consumes two result topics
and attempts to diff them. Later this should actually compare
the results of each subscription individually to verify that every
result is the same. For the intial version, I kept it simpler. All it does
is compare the count of subscriptions results it got on each topic for
a given scheduled timestamp, and report the difference.