feat: Add app_id field and propogate to Datadog (#2551)
Add the app_id field to the request and propagate it through to the attribution
logging in the querylog module.

Added an attribution module to handle all the attribution specific code, since
this logic will get more complex going forward. Ideally this module is
responsible for tracking everything attribution related so the query pipeline
code can remain largely unchanged. Moved the attribution logging logic from the
querylog file into the attribution module so it can be expanded later.

Created a default file for all the AppIDs. Ultimately this will be augmented by
files stored in ops. There is a default AppID as a placeholder, and an invalid
AppID to make tracking invalid cases a little easier.

This will also stop logging the team/feature/parent_api to Datadog. They are
deprecated and no one is using that data so it is safe to remove.