feat(debt) Refactoring consumer step 2: Use ConsumerStrategyFactory for the multi storage consumer (#2569)
There is quite a bit of duplication between ConsumerStrategyFactory and MultistorageConsumerProcessingStrategyFactory. Specifically in how the collector is built.

The ultimate goal is to have the same pipeline between multistorage consumer and standard consumer. The multistorage consumer is strictly a superset of the standard consumer (with one storage only). In order to get there I am trying to remove the duplications step by step then I will merge the steps that are not strictly duplicated.

This step moves in the direction of using the same factory instead of having a custom multistorage factory MultistorageConsumerProcessingStrategyFactory. I cannot do that in one step because MultistorageConsumerProcessingStrategyFactory has a transformation step before the filtering and that is not supported by ConsumerStrategyFactory. So it creates the transform step and delegates the rest to ConsumerStrategyFactory.

It also extracts MultistorageConsumerProcessingStrategyFactory methods so that they will be easier to merge with the single consumer ones.