fix: Change column entity validator to WARN by default (#2577)
We are not currently validating that columns referenced in queries are actually
available on the Entity. There was an initiative to do those checks, but it
stalled. This make any missing column errors throw a Sentry warning so we can
see how often our validators would catch broken queries. If things are working
as expected, we can turn the validators on by default and remove the xfail
from the test.