chore(retention): Fix errors and transactions retention period
When we receive messages from upstream, we are not really sure what
values are being sent for retention days. That means it is possible
that we could have any number for retention days. That is bad since
our partitions are by retention days and timestamps.

The change buckets the retention days to either 30 or 90 by using
the ceiling of the retention days parameter passed to us.