feat(subscriptions): Count the off by one results separately in the verifier (#2607)
A number of results seem to be off by one in the new pipeline + Tiger vs the old
subscriptions pipeline + main ClickHouse cluster. Since there's not a great deal
we can do about these, this change counts those separately and does not log them
to Sentry so we get more visibility on those results that might be more significantly different.