fix(subscriptions): Fix InvalidRangeError when rebalancing occurs
This fixes SNUBA-2R0.

We weren't properly resetting ticks and offsets stored internally in the
CommitLogTickConsumer on partition revocation. This made it possible for the
consumer to receive an earlier tick (after revocation and reassignment) and
attempt to construct an interval with the last one it saw before the revocation
happened, which is invalid. It also means it was possible that scheduling for
a tick was missed on rebalancing / deployment.