fix(subscriptions): Fix InvalidRangeError (again)
https://github.com/getsentry/snuba/pull/2611 wasn't quite right
since the partitions we are keeping track of in self.__previous_messages
actually refer to the partitions on the original topic not the commit log
topic which is the one we are consuming from. We need to completely clear
self.__previous messages on revocation to avoid building invalid intervals
or skipping ticks during rebalance.