feat: Support separate transactions topic in Snuba
Add the `transactions` and `snuba-transactions-commit-log` topics
to Snuba in order to support moving these to a separate topic entirely.
Currently these should still be mapped to the `events` and `snuba-commit-log`
topics in Snuba settings in every environment as they still share a topic.
Once we are ready to start producing to the new topic, we can remove the
mapping in settings.