fix(subscriptions): Flush producer on rebalance (#2648)
Bringing this experiment back since https://github.com/getsentry/snuba/pull/2644
didn't work.

This is an attempt to fix an error we are seeing in the executor logs
about the producer terminating with unflushed messages still in queue.

Now we manually call flush() when the strategy is closed. For now this
is just a test - if it actually works, then we will update Arroyo to properly
expose the flush() method instead of relying on this hack.