ref(tiger): Allow increasing cache wait time for tiger (#2651)
We are noticing an increase in ExecutionTimeoutError on the tiger
cluster. In order to debug what's happening, I would like the ability
to increase the cache wait timeout period for tiger clusters. The max
query execution time for a clickhouse query would be still 30 seconds.
But I would increase the wait timeout to see if the result gets
populated with some additional wait time.