feat(subscriptions): Combined scheduler and executor (#2666)
An alternative to the separate scheduler and executor that can both schedule and execute subscriptions for multiple entities in the same dataset in one process. Designed for use cases where the number of subscriptions to be run is not high and we want to reduce the number of containers that are run (e.g. dev and self-hosted).