fix: Allow arbitrary AppIDs (#2692)
This was originally written with the intention of having a registry of AppIDs
that could be used to check and validate incoming values. It's not clear if that
is still a strategy we want to use in the future, so for now allow arbitrary
values so we can start collecting some basic data.