fix(rate-limits): Create copy of request settings (#2733)
The current implementation of RateLimiterDelegate uses the same
RequestSettings passed to it in the constructor. That causes a bug
since the original rate_limit_params of HTTPRequestSettings are stored
in a list. The same list gets used by both the threads processing the
query.

The end result of this is that instead of processing 3 sets of rate
limits (project, table, global), secondary queries are actually getting 6
sets of rate limits (all of the above multiplied by 2). This leads
to overcounting since a query run in a single execution pipeline
gets counted in both execution pipelines.

To fix this we use a deep copy of the RequestSettings.