feat(subscriptions): Record information about size of each tick buffer
We experienced an incident where some partitions heavily lagged
behind others. There was not a lot of visibility about the
contents of the TickBuffer and how long it was waiting for ticks
to be able to be submitted to the next step. This is important for
datasets that run in "global" mode, as subscriptions on those datasets
wait for all partitions to reach a timestamp before they can be executed.