fix(subscriptions): Commit the right offset in the scheduler and executor (#2746)
The offset to be committed should always be the *next_offset* - the minimum
offset of the next message to be processed (which is one higher than the
offset of the last successfully processed message). Previously we
were committing the offset of the current message which is one offset behind
what it should be.